Rewrite for manifest v3
This patch rewrites the extension quite extensively.
Background page is now an event page which will be suspended when not in
use. All update operations still happen in the background-page context,
the UI communicates with the background using port based connection.

Functionally this patch doesn't change anything, except to inform the
user if update process did not change any bookmarks.